Just testing some minor changes

+++ ZombieSurvival/src/gs/tnt/dev/minecraft/ZombieSurvival/Variables.java
@@ -11,9 +11,9 @@ public final class Variables
 {
 	public static final String pluginName = "ZombieSurvival";
 	public static final String pluginVersion = "0.1.0.0";
-	public static final int pluginVersionMajor = 0;
-	public static final int pluginVersionMinor = 1;
-	public static final int pluginVersionBuild = 0;
-	public static final int pluginVersionRevision = 0;
+	public static final byte pluginVersionMajor = 0;
+	public static final byte pluginVersionMinor = 1;
+	public static final byte pluginVersionBuild = 0;
+	public static final byte pluginVersionRevision = 0;
 	
 }
